Refactor how sync resolves
This first checks if anything is in the queue or buffer to be processed. If so, then it waits for a `replicate.progress` event (where the progress is complete) and then waits for the `replicated` event.

The downside to this is that if other things are being added to the queue rapidly, then this may hang indefinitely. That said, currently as this is used in orbit-db, this would not block anything other than the new event I have proposed adding, so it should be fine?

Ideally, what should really happen is that we should have a way of knowing when the items derived from the given heads have been replicated. Currently, however, the queue is a black hole and there's no way to track the graph that branches out from the heads. The queue would likely need to be refactored for this approach.